refactor(tui): move composer behavior into app/composer.rs (#3314 step 2)
Pure code motion for the App god-object split:

- app/composer.rs gains an inherent impl ComposerState holding the 22
  composer-only methods that touch no App-level state (cursor/selection
  math, history-search queries, paste scrubbing, draft recovery,
  vim mode predicates). vim_is_normal_mode/vim_is_visual_mode drop the
  now-redundant 'self.composer.' qualifier, the only textual adjustment.
- An extension impl App in the same module keeps the 60 composer methods
  that also need App-level state (needs_redraw, status_message, clipboard,
  workspace, submit/consolidate paths) out of the central impl block.
- File-scope paste/scrub helpers move with them: ComposerHistorySearch,
  InputHistoryDraft, char_count, byte_index_at_char, remove_char_at,
  remove_char_range, prev/next_grapheme_boundary, normalize_paste_text,
  sanitize_api_key_text, the raw-mouse-report/control-sequence scrubbers,
  and the MAX_SUBMITTED_INPUT_CHARS / MAX_COMPOSER_DISPLAY_CHARS /
  MAX_DRAFT_HISTORY caps.
- app.rs re-exports keep every existing path stable
  (crate::tui::app::char_count, ComposerHistorySearch, etc.); the three
  helpers still needed by the onboarding API-key input methods that remain
  in app.rs (byte_index_at_char, remove_char_at, sanitize_api_key_text)
  become pub(crate), the minimal visibility the split requires.
- The Deref/DerefMut bridge to ComposerState on App is untouched; no call
  site changes, no behavior change.
